tva et ref

+++ client/app/devis/dupliquerdevis.component.ts
@@ -320,7 +320,7 @@ export class DupliquerDevisComponent implements OnInit {
         let total = 0;
         for (let produit of this.produitDevis) {
             if (produit.taux == 2.1) {
-                total += produit.qte_devis * produit.prix_devis * (parseInt(produit.taux) / 100);
+                total += produit.qte_devis * produit.prix_devis * (parseFloat(produit.taux) / 100);
             }
         }
         return total;
@@ -331,7 +331,7 @@ export class DupliquerDevisComponent implements OnInit {
         let total = 0;
         for (let produit of this.produitDevis) {
             if (produit.taux == 5.5) {
-                total += produit.qte_devis * produit.prix_devis * (parseInt(produit.taux) / 100);
+                total += produit.qte_devis * produit.prix_devis * (parseFloat(produit.taux) / 100);
             }
         }
         return total;
@@ -373,7 +373,7 @@ export class DupliquerDevisComponent implements OnInit {
         let total = 0;
         for (let produit of this.produitDevisOptions) {
             if (produit.taux == 2.1) {
-                total += (parseInt(produit.taux) / 100) * produit.prix_devis * produit.qte_devis;
+                total += (parseFloat(produit.taux) / 100) * produit.prix_devis * produit.qte_devis;
             }
         }
         return total;
@@ -383,7 +383,7 @@ export class DupliquerDevisComponent implements OnInit {
         let total = 0;
         for (let produit of this.produitDevisOptions) {
             if (produit.taux == 5.5) {
-                total += (parseInt(produit.taux) / 100) * produit.prix_devis * produit.qte_devis;
+                total += (parseFloat(produit.taux) / 100) * produit.prix_devis * produit.qte_devis;
             }
         }
         return total;
